requirements.txt
beautifulsoup4==4.5.3
requests==2.13.0
djangorestframework==3.5.4
Django==1.10.5